+++ pom.xml
@@ -5,10 +5,10 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.vaadin.componentfactory</groupId>
-    <artifactId>date-range-picker-parent</artifactId>
-    <version>5.0.3-SNAPSHOT</version>
+    <artifactId>date-time-range-picker-parent</artifactId>
+    <version>1.0.0-SNAPSHOT</version>
     <packaging>pom</packaging>
-    <name>Date Range Picker Parent</name>
+    <name>Date Time Range Picker Parent</name>
 
     <organization>
         <name>Vaadin Ltd</name>
@@ -36,7 +36,7 @@
     </licenses>
 
     <modules>
-        <module>vcf-date-range-picker</module>
-        <module>vcf-date-range-picker-demo</module>
+        <module>vcf-date-time-range-picker</module>
+        <module>vcf-date-time-range-picker-demo</module>
     </modules>
 </project>
