+++ vcf-date-time-range-picker/assembly/MANIFEST.mf
